refactor: move DNS tests to DnsTests.cs and use List in GetServers()
- Move GetServers/GetEndPoints DNS topology tests into existing DnsTests
  class with DnsTestsFixture (add DnsClusterServer to the fixture).
- Use List<ValkeyServer> instead of array+index in GetServers().
- Add host parameter to StandaloneServer for consistency.
- Make TLS server startup non-fatal in DnsTestsFixture so non-TLS
  tests can still run when TLS servers fail to start.
- Remove separate DnsGetServersTests.cs file.
